motion-of-no-confidence.md: Put teeth into the committee responsibilities
I don't expect us to need these teeth, but there's not much point in
having a list of responsibilities without (potential) consequences for
violating them.  This gives both the steering committee and the
Foundation membership at large the ability to force early elections if
they feel that the a member or members of the current committee are
neglecting their responsibilities.

I shortened the usual 90-day lead time for the early elections,
because that seemed to be too large a fraction of the annual term.
Between 45 and 60 days should provide both a sufficiently long period
for campaigning and a sufficiently short period for minimizing further
negligence.