motion-of-no-confidence.md: Drop 'unlikely'
In the introductory paragraph, I had been trying to imply that I think
no-confidence motions won't happen frequently.  Katy Huff felt like it
was unnecessary noise:

On Thu, Mar 12, 2015 at 12:51:20PM -0700, Katy Huff wrote [1]:
> Necessary to say unlikely? Let's stick to the facts in this document
> as much as possible.

and it is.  This commit drops the word, and leaves it to the readers
to draw their own conclusions about the expected frequency of the
motion.  For a facts-based approach we should be recording any of
these events (if we ever get one) in our election results (although I
don't see a place for those at the moment ;).

[1]: https://github.com/swcarpentry/board/pull/24/files#r26337801