executive-director: Put in charge of SC resources
E.g. this makes it more obvious the the executive branch is in charge
of moderating discussion lists as it sees fit.  These responsibilities
don't talk about delegation vs. direct action by the Executive
Director, but I expect delegation is fine as long as the buck stops
with the ED.