executive-director: Explicitly allow delegated power
But you can't delegate responsibility (e.g. the SC can boot the ED for
a proxy-ED action as if the ED had acted directly).  This borrows
proxy wording from proxy-membership.md, and helps avoid the false
assumption that the ED must carry out all executive action personally.

I'm not sure how the funding will work out for paid proxies (e.g. if a
community manager is hired).  I'd guess the ED is free to do that
however they like, but will have to requisition funds from the SC.
And as a condition on supplying the funds, the SC may require the ED
to provide a job description, salary estimate, etc., etc.  So [1].
There's not a lot of wording to that effect in the current docs, so
I've left it out for now.

[1]: https://en.wikipedia.org/wiki/Power_of_the_purse