subcommittees: Restructure for per-subcommittee directories
Also for task forces.  This streamlines the generic docs and listings,
and pushes the per-subcomittee (or task force) details down into their
own separate area.  With a specific area, the per-subcomittee
information may be more likely to stay up to date (by avoiding
diffusion of responsibility [1]) and there is more space for useful
information like links to blogs, mailing lists, repositories, etc.

I've attempted to fill in GitHub usernames and some of this auxiliary
project information to the best of my knowledge, but it's a lot of
shifting around information I'm not particularly familiar with, so I
may have gotten some things wrong.

[1]: https://en.wikipedia.org/wiki/Diffusion_of_responsibility