subcommittees/mentoring/README: Remove "past year" from debriefing paragraph
The old wording is stale now that it's 2017.  The new wording makes
the paragraph more compact by merging the previously-doubled links to
the same target.  I also removed the references to "blog", because the
important idea is "follow this link for information about these
sessions"; whether that information is on a blog or not isn't very
important.